Add initial config backend
This adds a new  class which uses __setattr__ and __getattr__ to wrap
TinyDB so that configuration backed by disk can be set and retrieved
via properties on the config object.

Values of the configuration are also type checked at runtime to prevent
setting the wrong values. The configuration types exist as type hints
at the top of the Config class.

The config builtin was also added, although this should likely be moved
to a module so it can be used via the CLI as well. If the config builtin
is called with one arg it will dump the current configuration, if it is
called with one it will print that value, if it is called with two or
more it will treat the first as the config key and the rest as the
values.